+++ lazy_errors/src/surrogate_error_trait/prelude.rs
@@ -46,7 +46,9 @@
 //! [_inner error type_ `I`]: crate::Error#inner-error-type-i
 //! [CUSTOM]: crate#example-custom-error-types
 
-pub use crate::{err, try2, OrCreateStash, OrStash, OrWrap, OrWrapWith};
+pub use crate::{
+    err, try2, OrCreateStash, OrStash, OrWrap, OrWrapWith, StashErr,
+};
 
 /// Type alias for [`crate::StashedResult`]
 /// to use a boxed [_inner error type_ `I`](crate::Error#inner-error-type-i),